yaicAbgi
Yaicabgi bison yacc is a parser generator tool, commonly used in the field of compiler construction. It is a part of the GNU Project and is based on the well-known tool, Bison (Yacc is another name for Bison). The provided link (https://www.gnu.org/software/bison/manual/bison.html) directs to the official documentation of Bison, which includes a comprehensive manual on using Bison for creating parsers.